requirements.txt
# Google APIs
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Gemini AI (new package; google-generativeai is deprecated)
google-genai>=1.0.0

# Utilities
python-dateutil>=2.8.2
python-dotenv>=1.0.0
